Feature/angular java user integration

+++ src/main/java/com/javauserbackend/UserRepository.java
@@ -2,6 +2,7 @@ package com.javauserbackend;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-// This interface extends JpaRepository to provide CRUD operations for the User entity.
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
+
+
